Move `FlashFrequency` and `FlashMode` enums into the flasher module, where they belong
