+++ espflash/src/flasher.rs
@@ -9,7 +9,7 @@ use crate::{
     chip::Chip,
     command::{Command, CommandType},
     connection::Connection,
-    elf::{ElfFirmwareImage, FirmwareImage, FlashFrequency, FlashMode, RomSegment},
+    elf::{ElfFirmwareImage, FirmwareImage, RomSegment},
     error::{ConnectionError, FlashDetectError, ResultExt},
     image_format::ImageFormatId,
     stubs::FlashStub,
@@ -27,6 +27,81 @@ const FLASH_SECTORS_PER_BLOCK: usize = FLASH_SECTOR_SIZE / FLASH_BLOCK_SIZE;
 const CHIP_DETECT_MAGIC_REG_ADDR: u32 = 0x40001000;
 
 const EXPECTED_STUB_HANDSHAKE: &str = "OHAI";
+
+#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Display, EnumVariantNames)]
+#[repr(u8)]
+pub enum FlashFrequency {
+    #[strum(serialize = "12M")]
+    Flash12M,
+    #[strum(serialize = "15M")]
+    Flash15M,
+    #[strum(serialize = "16M")]
+    Flash16M,
+    #[strum(serialize = "20M")]
+    Flash20M,
+    #[strum(serialize = "24M")]
+    Flash24M,
+    #[strum(serialize = "26M")]
+    Flash26M,
+    #[strum(serialize = "30M")]
+    Flash30M,
+    #[strum(serialize = "40M")]
+    Flash40M,
+    #[strum(serialize = "48M")]
+    Flash48M,
+    #[strum(serialize = "60M")]
+    Flash60M,
+    #[strum(serialize = "80M")]
+    Flash80M,
+}
+
+impl FromStr for FlashFrequency {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        use FlashFrequency::*;
+
+        match s.to_uppercase().as_str() {
+            "12M" => Ok(Flash12M),
+            "15M" => Ok(Flash15M),
+            "16M" => Ok(Flash16M),
+            "20M" => Ok(Flash20M),
+            "24M" => Ok(Flash24M),
+            "26M" => Ok(Flash26M),
+            "30M" => Ok(Flash30M),
+            "40M" => Ok(Flash40M),
+            "48M" => Ok(Flash48M),
+            "60M" => Ok(Flash60M),
+            "80M" => Ok(Flash80M),
+            _ => Err(Error::InvalidFlashFrequency(s.to_string())),
+        }
+    }
+}
+
+#[derive(Copy, Clone, Debug, EnumVariantNames)]
+#[strum(serialize_all = "UPPERCASE")]
+pub enum FlashMode {
+    Qio,
+    Qout,
+    Dio,
+    Dout,
+}
+
+impl FromStr for FlashMode {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let mode = match s.to_uppercase().as_str() {
+            "QIO" => FlashMode::Qio,
+            "QOUT" => FlashMode::Qout,
+            "DIO" => FlashMode::Dio,
+            "DOUT" => FlashMode::Dout,
+            _ => return Err(Error::InvalidFlashMode(s.to_string())),
+        };
+
+        Ok(mode)
+    }
+}
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Display, EnumVariantNames)]
 #[repr(u8)]
