+++ espflash/src/image_format/mod.rs
@@ -5,11 +5,7 @@ use serde::Deserialize;
 use strum_macros::{Display, EnumVariantNames, IntoStaticStr};
 
 pub use self::{esp32bootloader::*, esp32directboot::*, esp8266::*};
-use crate::{
-    chip::Chip,
-    elf::{FlashFrequency, RomSegment},
-    error::Error,
-};
+use crate::{chip::Chip, elf::RomSegment, error::Error, flasher::FlashFrequency};
 
 mod esp32bootloader;
 mod esp32directboot;
